Warn about transitive dependencies
This commit change the way how namcap handles dependencies to meet with our recent guideline to not rely on transitive dependencies, as they might break, if one of the dependencies is updated.[1]

Previously, namcap warned about dependencies if they are included but already satisfied, even if these are transitive dependencies. From now, namcap will warn if a dependency is not explicitly specified, and won't show a warning if these are already satisfied by other dependencies.

This commit also simplifies and fixes various issues in the dependency handling logic, e.g.:
- Strip the version numbers when searching what provides a package, otherwise it doesn't match any.
- Remove some slow recursive dependency search. With the new logic we only need it in case of the explicitly specified dependencies.
- Always show at least an info message when a dependency satisfied. Useful to see which components need a library.

[1] https://wiki.archlinux.org/title/Arch_package_guidelines#Package_dependencies