Fix tests
core/binutils switched from DT_RPATH to DT_RUNPATH for --rpath since
2.39-4 [1][2], and thus the rpath rule cannot detect it. I force old
DT_RPATH here for testing.

There is another runpath rule for DT_RUNPATH tags. It probably needs a
test, but that's out of scope for this commit.

As a side note, here is why binutils 2.39-4 changes the behavior of ld
(excerpted from [1]):

> The --enable-new-dtags configure option causes the linker to set the
> --enable-new-dtags option as default. This means that "-Wl,-rpath"
> will produce the DT_RUNPATH symbol instead of DT_RPATH

[1] https://bugs.archlinux.org/task/76274
[2] https://github.com/archlinux/svntogit-packages/commit/9c3a2d6851aa2bceeb44f33499c3c19a519ca16d