Add pcdepends rule
This rule checks if dependencies from pkg-config .pc files are satisfied. These are used by the 'pkg-config --cflags --libs <library name>' command to set the appropriate CFLAGS and LDFLAGS when building any dependent (direct and indirect) packages. If any of dependencies specified in 'Requires' or 'Requires.private' fields are missing, pkg-config trows an error.
These dependencies are mostly covered by link-level dependencies, but there are some exceptions, this is why we need this rule.