refactor: Use raw strings for regular expressions
Taking a hint from pyright which notes '\s' is a valid regular
expression but *not* a valid string literal, I just made all the re
input strings raw. Some had '\t' and other ambiguous usage but this
should be clearer in the end.

Exactly one exsting line of code had a correctly escaped backslash in
a regex, which I converted to raw.