feat: Use SPDX license identifiers when matching package licenses
Rely on *license-expression* to take care of parsing license strings.
Compare license identifiers with all known identifiers (licenses and
exceptions).

Derive all known SPDX license identifiers from a name list, shipped with
the licenses package and add an error if the license identifier can not
be found.
Remove the use of `special_licenses` (e.g. BSD, MIT, etc.) as "bad
list", because with the new SPDX license identifiers we maintain an
extensive "good list" instead.

This follows the specification of
https://rfc.archlinux.page/0016-spdx-license-identifiers/
which outlines the use of all plaintext SPDX license files for the
purpose of a new common licenses package.

Signed-off-by: David Runge <dvzrv@archlinux.org>